+++ package/scripts/lib.js
@@ -1,8 +1,12 @@
 import * as cp from 'node:child_process';
 import * as path from 'node:path';
 import * as fs from 'node:fs/promises';
-
 import { readFileSync } from 'node:fs';
+
+import babel from '@rollup/plugin-babel';
+import nodeResolve from '@rollup/plugin-node-resolve';
+import commonJs from '@rollup/plugin-commonjs';
+import terser from '@rollup/plugin-terser';
 
 export const packageFolder = path.join(__dirname, '..');
 export const pkg = JSON.parse(readFileSync(path.resolve(packageFolder, 'package.json')).toString());
@@ -39,3 +43,41 @@ export async function copyRunTimeDepsToOut(target) {
 }
 
 export const extensions = ['.js', '.ts'];
+
+
+const entryPointsAMD = ['kustoMode', 'kustoWorker', 'monaco.contribution'];
+
+/**
+ * Non-dev builds are minified, bundled, and transpiled so they can be consumed
+ * directly, without the consumers running their own bundler or transpiler.
+ *
+ * @param {'dev' | 'min'} type
+ * @returns {import('rollup').RollupOptions}
+ */
+export function rollupAMDConfig(type) {
+    return {
+        external: ['monaco-editor-core'],
+        input: Object.fromEntries(entryPointsAMD.map((e) => [e, path.join(packageFolder, 'src', e + '.ts')])),
+        plugins: [
+            nodeResolve({ extensions }),
+            commonJs(),
+            babel({
+                extensions,
+                babelHelpers: 'inline',
+                presets: [['@babel/preset-env', { targets: { ie: 11 } }], '@babel/preset-typescript'],
+            }),
+            type === 'min' && terser(),
+        ],
+        output: {
+            name: 'test',
+            banner,
+            format: 'amd',
+            amd: { autoId: true, basePath: 'vs/language/kusto' },
+            dir: path.join(packageFolder, 'release', type),
+            globals: {
+                'monaco-editor-core': 'monaco',
+            },
+            sourcemap: !process.env.CI,
+        },
+    };
+}
