--- text.txt
@@ -1,1 +1,0 @@
-posted in Kusto Web Explorer / General at February 9, 2023 6:52:54 PM PST In case you missed it 👆 What's the email of copy editor?
